app/core/token.py
from jose import jwt
from datetime import datetime, timedelta
from typing import Optional

SECRET_KEY = "3b3f6a5e4863483d88b1d49b155d1961c05711e729f8eae86e0647f3b8e77d6e"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 60


def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -> str:
    to_encode = data.copy()
    expire = datetime.utcnow() + (expires_delta or timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES))
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
