requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
mysql-connector-python
alembic
pydantic-settings
